Refactor transport and serialize_parse (now marshalling) headers
