+++ src/apps/zeromq/terminate/terminate.cpp
@@ -23,8 +23,8 @@
 #include <signal.h>
 #include <sys/types.h>
 
+#include "goby/middleware/marshalling/protobuf.h"
 #include "goby/middleware/protobuf/terminate.pb.h"
-#include "goby/middleware/serialize_parse_protobuf.h"
 #include "goby/middleware/terminate/groups.h"
 #include "goby/zeromq/protobuf/terminate_config.pb.h"
 #include "goby/zeromq/single-thread-application.h"
