+++ src/middleware/single-thread-application.h
@@ -27,8 +27,8 @@
 
 #include "goby/middleware/application.h"
 #include "goby/middleware/thread.h"
-#include "goby/middleware/transport-interprocess.h"
-#include "goby/middleware/transport-intervehicle.h"
+#include "goby/middleware/transport/interprocess.h"
+#include "goby/middleware/transport/intervehicle.h"
 
 #include "goby/middleware/terminate/terminate.h"
