+++ src/moos/middleware/moos_plugin_translator.cpp
@@ -20,7 +20,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Goby.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "goby/middleware/serialize_parse_protobuf.h"
+#include "goby/middleware/marshalling/protobuf.h"
 
 #include "moos_plugin_translator.h"
 #include "goby/moos/moos_translator.h"
