+++ src/test/zeromq/middleware_interprocess_forwarder/test.cpp
@@ -25,9 +25,9 @@
 #include <atomic>
 #include <deque>
 
-#include "goby/middleware/serialize_parse_protobuf.h"
+#include "goby/middleware/marshalling/protobuf.h"
 #include "goby/middleware/transport.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 #include "goby/util/debug_logger.h"
 #include "test.pb.h"
