+++ src/test/zeromq/middleware_regex/test.cpp
@@ -25,10 +25,10 @@
 #include <atomic>
 #include <deque>
 
-#include "goby/middleware/serialize_parse_protobuf.h"
+#include "goby/middleware/marshalling/protobuf.h"
 #include "goby/middleware/transport.h"
 #include "goby/util/debug_logger.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 #include "test.pb.h"
