+++ src/test/zeromq/middleware_speed/test.cpp
@@ -27,10 +27,10 @@
 
 #include <boost/units/io.hpp>
 
-#include "goby/middleware/serialize_parse_protobuf.h"
+#include "goby/middleware/marshalling/protobuf.h"
 #include "goby/middleware/transport.h"
 #include "goby/util/debug_logger.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 #include "test.pb.h"
