+++ src/test/zeromq/zeromq_and_intervehicle/test.cpp
@@ -25,11 +25,11 @@
 #include <atomic>
 #include <deque>
 
-#include "goby/middleware/serialize_parse_dccl.h"
-#include "goby/middleware/serialize_parse_protobuf.h"
+#include "goby/middleware/marshalling/dccl.h"
+#include "goby/middleware/marshalling/protobuf.h"
 #include "goby/middleware/transport.h"
 #include "goby/util/debug_logger.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 #include "test.pb.h"
