+++ src/zeromq/multi-thread-application.h
@@ -24,7 +24,7 @@
 #define ZEROMQ_MULTI_THREAD_APPLICATION20190605H
 
 #include "goby/middleware/multi-thread-application.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 namespace goby
 {
