+++ src/zeromq/single-thread-application.h
@@ -24,7 +24,7 @@
 #define ZEROMQ_SINGLE_THREAD_APPLICATION20190605H
 
 #include "goby/middleware/single-thread-application.h"
-#include "goby/zeromq/transport-interprocess.h"
+#include "goby/zeromq/transport/interprocess.h"
 
 namespace goby
 {
